move test files into public folder
