+++ sdk/communication/communication-identity/test/public/utils/mockHttpClients.ts
@@ -2,8 +2,8 @@
 // Licensed under the MIT license.
 
 import { HttpClient, WebResourceLike, HttpOperationResponse, HttpHeaders } from "@azure/core-http";
-import { CommunicationAccessToken } from "../../src";
-import { CommunicationIdentityAccessTokenResult } from "../../src/generated/src/models";
+import { CommunicationAccessToken } from "../../../src";
+import { CommunicationIdentityAccessTokenResult } from "../../../src/generated/src/models";
 
 export const createMockHttpClient = <T = Record<string, unknown>>(
   status: number = 200,
